requirements.txt
beautifulsoup4==4.6.3
boto3==1.9.0
botocore==1.12.0
certifi==2018.8.24
chardet==3.0.4
cssselect==1.0.3
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
jieba3k==0.35.1
jmespath==0.9.3
lxml==4.2.4
newspaper3k==0.2.7
nltk==3.3
numpy==1.15.1
Pillow==5.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
s3transfer==0.1.13
six==1.11.0
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.23
